Mini refactor on ai naming to ollamaAPI
